Updated workflow_metric_reporter.py
Added two methods to workflow_metric_report.py
